Add Computer factory
